requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
assemblyai==0.35.1
beautifulsoup4==4.12.3
blinker==1.9.0
Bootstrap-Flask==2.3.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docopt==0.6.2
Flask==2.3.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.0
Flask-WTF==1.1.1
googletrans-modified==1.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.25.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
Js2Py==0.74
Mako==1.3.6
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.1.3
packaging==24.2
pipwin==0.5.2
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyjsparser==2.7.1
PyMySQL==1.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
requests==2.32.3
rfc3986==1.5.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==1.4.46
srt==3.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vosk==0.3.45
websockets==14.1
Werkzeug==3.1.3
WTForms==3.2.1
